rework action types
