+++ src/shared/store/actionTypes/products.ts
@@ -1,0 +1,21 @@
+export enum PRODUCT_ACTIONS {
+  REQUEST_FETCH_PRODUCTS = "REQUEST_FETCH_PRODUCTS",
+  FETCH_PRODUCTS_PENDING = "FETCH_PRODUCTS_PENDING",
+  FETCH_PRODUCTS_SUCCESS = "FETCH_PRODUCTS_SUCCESS",
+  FETCH_PRODUCTS_FAILED = "FETCH_PRODUCTS_FAILED",
+
+  REQUEST_ADD_PRODUCT = "REQUEST_ADD_PRODUCT",
+  ADD_PRODUCT_PENDING = "ADD_PRODUCT_PENDING",
+  ADD_PRODUCT_SUCCESS = "ADD_PRODUCT_SUCCESS",
+  ADD_PRODUCT_FAILED = "ADD_PRODUCT_FAILED",
+
+  REQUEST_EDIT_PRODUCT = "REQUEST_EDIT_PRODUCT",
+  EDIT_PRODUCT_PENDING = "EDIT_PRODUCT_PENDING",
+  EDIT_PRODUCT_SUCCESS = "EDIT_PRODUCT_SUCCESS",
+  EDIT_PRODUCT_FAILED = "EDIT_PRODUCT_FAILED",
+
+  REQUEST_DELETE_PRODUCT = "REQUEST_DELETE_PRODUCT",
+  DELETE_PRODUCT_PENDING = "DELETE_PRODUCT_PENDING",
+  DELETE_PRODUCT_SUCCESS = "DELETE_PRODUCT_SUCCESS",
+  DELETE_PRODUCT_FAILED = "DELETE_PRODUCT_FAILED",
+}
