+++ src/shared/store/reducers/products.ts
@@ -1,16 +1,13 @@
 import { IProduct } from "shared/models";
-import { PRODUCT_ACTIONS } from "../actions/productActions";
+import * as productActions from "../actions/products";
+import { PRODUCT_ACTIONS } from "../actionTypes/products";
+import { ActionTypes } from "../types";
 
 interface ProductReducerState {
   products: IProduct[];
   isLoading: boolean;
   isError: boolean;
   error: string;
-}
-
-interface ProductAction {
-  type: PRODUCT_ACTIONS;
-  payload?: IProduct[] | IProduct | string | boolean;
 }
 
 const defaultState: ProductReducerState = {
@@ -22,7 +19,7 @@ const defaultState: ProductReducerState = {
 
 export const productReducer = (
   state: ProductReducerState = defaultState,
-  action: ProductAction
+  action: ActionTypes<typeof productActions>
 ): ProductReducerState => {
   switch (action.type) {
     case PRODUCT_ACTIONS.FETCH_PRODUCTS_PENDING:
@@ -36,14 +33,14 @@ export const productReducer = (
         ...state,
         isLoading: false,
         isError: false,
-        products: action.payload as IProduct[],
+        products: action.payload,
       };
     case PRODUCT_ACTIONS.FETCH_PRODUCTS_FAILED:
       return {
         ...state,
         isLoading: false,
         isError: true,
-        error: action.payload as string,
+        error: action.payload,
       };
     case PRODUCT_ACTIONS.ADD_PRODUCT_PENDING:
       return {
@@ -62,7 +59,7 @@ export const productReducer = (
         ...state,
         isLoading: false,
         isError: true,
-        error: action.payload as string,
+        error: action.payload,
       };
     case PRODUCT_ACTIONS.EDIT_PRODUCT_PENDING:
       return {
