+++ src/shared/store/saga/index.ts
@@ -1,5 +1,5 @@
 import { fork } from "redux-saga/effects";
-import { productWatcher } from "./productSaga";
+import { productWatcher } from "./products";
 
 export function* rootSaga() {
   yield fork(productWatcher);
